Limit the Rhelp commands search paths
The `package` argument in the call to the R help() function was set to
NULL in order to limit the number of packages searched for a help file.
This change both speeds up the Rhelp command and makes sure that help
files for base R functions are returned rather than functions with the
same name in a third party package.