Remove option to open help files inside vim
The ability to open R help files in a vim buffer was removed. If
`g:rook_help_type = 'text'`, the text based help file is now opened
inside of the target R console.

The opening of help files inside vim never worked properly. In
particular, if the same function existed in multiple packages, there was
no mechanisms to choose which one to use without specifically
specifying the package in the Rhelp call. By opening the text based help
file inside the R console, the interactive picker can be used to select
which package to show help for a particular function.